refactor: enhance layout resposivness
